# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: images
    description: Image upload and management
  - name: forms
    description: Form submissions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /images:
    get:
      operationId: listImages
      tags: [images]
      summary: List all uploaded images
      responses:
        "200":
          description: List of images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImageRecord"
  /images/upload:
    post:
      operationId: uploadImage
      tags: [images]
      summary: Upload an image to Cloudinary with auto-enhancement
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                label:
                  type: string
              required:
                - file
                - label
      responses:
        "200":
          description: Uploaded image info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageRecord"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /images/{id}:
    delete:
      operationId: deleteImage
      tags: [images]
      summary: Delete an image record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /forms/inquire:
    post:
      operationId: submitInquiry
      tags: [forms]
      summary: Submit an inquiry form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InquiryForm"
      responses:
        "200":
          description: Inquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /forms/visit:
    post:
      operationId: submitVisit
      tags: [forms]
      summary: Schedule a campus visit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitForm"
      responses:
        "200":
          description: Visit scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /forms/inquiries:
    get:
      operationId: listInquiries
      tags: [forms]
      summary: List all inquiry submissions
      responses:
        "200":
          description: List of inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InquiryRecord"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ImageRecord:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        cloudinaryUrl:
          type: string
        publicId:
          type: string
        originalFilename:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - label
        - cloudinaryUrl
        - publicId
        - originalFilename
        - createdAt
    InquiryForm:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        gradeLevel:
          type: string
        studentName:
          type: string
        message:
          type: string
      required:
        - firstName
        - lastName
        - email
    VisitForm:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        preferredDate:
          type: string
        gradeLevel:
          type: string
      required:
        - firstName
        - lastName
        - email
    InquiryRecord:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        gradeLevel:
          type: string
        studentName:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - email
        - createdAt
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success